rename httpdns_restful_demo into httpdns_api_demo
