+++ httpdns_api_demo/proguard-rules.pro
