+++ httpdns_api_demo/src/androidTest/java/alibaba/httpdns_api_demo/ApplicationTest.java
@@ -1,4 +1,4 @@
-package alibaba.httpdns_restful_demo;
+package alibaba.httpdns_api_demo;
 
 import android.app.Application;
 import android.test.ApplicationTestCase;
