+++ httpdns_api_demo/src/main/java/alibaba/httpdns_api_demo/NetworkRequestUsingHttpDNS.java
@@ -1,4 +1,4 @@
-package alibaba.httpdns_restful_demo;
+package alibaba.httpdns_api_demo;
 
 import android.content.Context;
 import android.os.Build;
