+++ httpdns_api_demo/src/main/res/layout/activity_main.xml
