+++ httpdns_api_demo/src/main/res/values/colors.xml
