+++ settings.gradle
@@ -1,1 +1,1 @@
-include ':oss_android_demo', ':man_android_demo', ':man_restful_demo', ':mac_android_demo', ':ots_android_demo', ':httpdns_restful_demo', ':httpdns_android_demo'
+include ':oss_android_demo', ':man_android_demo', ':man_restful_demo', ':mac_android_demo', ':ots_android_demo', ':httpdns_api_demo', ':httpdns_android_demo'
